requirements.txt
Flask>=2.0
python-jobspy
pandas>=1.0
gunicorn
